==> 05_severity_metaAnalysis.Rmd <==
---
output: html_document
editor_options: 
  chunk_output_type: inline
---

# Meta-analysis of powdery mildew severity

```{r MEsummary_Libraries2, message=FALSE, include=FALSE}
if (!require("pacman"))
   install.packages("pacman")
pacman::p_load(tidyverse,
               kableExtra,
               bomrang,
               RColorBrewer,
               metafor,
               here,
               netmeta,
               multcomp,
               dplyr,
               flextable)

if (!require("theme.usq"))
   devtools::install_github("adamhsparks/theme.usq")
library(theme.usq)
theme_set(theme_usq())
source(here("R/reportP.R"))

# Data
PM_dat_D <-
   read.csv("cache/PM_disease_clean_data.csv", stringsAsFactors = FALSE)
```



Inspect for a normal distribution of disease severity
```{r sev_hist}
hist(PM_dat_D$PM_final_severity)
```


Let's get started with the analysis by first finding the best model fit that answers our research question.

> When, in relation to PM establishes in the crop, should farmers begin spraying mungbean to obtain the highest fungicide efficacy for lowering disease severity: before PM establishing,  immediately after first sign of PM establishing, or after PM has become established in the crop. 

A secondary question to this aim is: does a second spray increase the efficacy?

To do this, in our model:

   - Powdery mildew severity 1-9 scale

   - Trial, which resolves combinations of categorical variables: year, location, row spacing, fungicide dose and cultivar; is set as a random intercept

### Define Trial
```{r define_trial}
PM_dat_D <-
   PM_dat_D %>%
   mutate(trial = paste(trial_ref,
                        year,
                        location,
                        host_genotype,
                        row_spacing,
                        sep = "_")) 
```


   - We will investigate the effect size of our treatment - spray management,


```{r Metafor-Sev-analysis}
PMsev_mv <- rma.mv(
   yi = PM_final_severity,
   vi,
   mods = ~ spray_management,
   method = "ML",
   random = list( ~ spray_management | trial, ~ 1 | id),
   struct = "UN",
   control = list(optimizer = "optim"),
   data = PM_dat_D
)

summary(PMsev_mv)
```

All spray schedules were effective at lowering the disease severity at the end of the season

```{r sevContrasts}
source("R/simple_summary.R") #function to provide a table that includes the treatment names in the contrasts

contrast_Ssum <-
   simple_summary(summary(glht(PMsev_mv, linfct = cbind(
      contrMat(rep(1, 6), type = "Tukey")
   )), test = adjusted("none")))

contrast_Ssum
```

```{r plotsevContrasts}
par(mar = c(5, 13, 4, 2) + 0.1)
plot(glht(PMsev_mv, linfct = cbind(contrMat(rep(
   1, 6
), type = "Tukey"))), yaxt = 'n')
axis(
   2,
   at = seq_along(contrast_Ssum$contrast),
   labels = rev(contrast_Ssum$contrast),
   las = 2,
   cex.axis = 0.8
)
```


```{r metafor_results}
# obtain number of treatments included in each moderator variable
k5 <-
   as.data.frame(table(PM_dat_D$spray_management)) %>%
   filter(Freq != 0) %>%
   pull(Freq)

k6 <-
   as.data.frame(table(PM_dat_D$trial_ref, PM_dat_D$spray_management)) %>%
   filter(Freq != 0) %>%
   group_by(Var2) %>%
   summarise(n()) %>%
   pull()

intercept <- round(PMsev_mv$b, 4)[1,1]

# create data.frame
results_mv <- data.frame(
   Moderator = c(
   "Intercept / No Spray control",
   "Early",
   "Late",
   "Late+",
   "Recommended",
   "Recommended+"
   ),
   N = k5,
   k = k6,
   Effect = round(PMsev_mv$b, 4),
   se = round(PMsev_mv$se, 4),
   CI_lower = round(PMsev_mv$ci.lb, 4),
   CI_upper = round(PMsev_mv$ci.ub, 4),
   z_val = round(PMsev_mv$zval, 4),
   p_val = reportP(PMsev_mv$pval, AsNumeric = FALSE, P_prefix = FALSE),
   eff_bar = round(1 -(intercept + PMsev_mv$b)/intercept,4)*100,
   eff_CI_high = round(1 -(intercept + PMsev_mv$ci.ub)/intercept,4)*100,
   eff_CI_low = round(1 -(intercept + PMsev_mv$ci.lb)/intercept,4)*100
   
)

# rename colnames to give table headings
colnames(results_mv)[c(1:4,6:12 )] <-
   c("Moderator",
     "N",
     "k",
     "mu",
     "CI_{L}",
     "CI_{U}",
     "Z",
     "P",
     "Efficacy",
     "Eff CI_lower",
     "Eff CI upper")

disease_estimates_table <-
   flextable(results_mv[c(2,5,6,3,4), c(1:4,6:7, 9:12)]) %>%
   align(j = 2:10, align = "center", part = "all") %>%
   fontsize(size = 8, part = "body") %>%
   fontsize(size = 10, part = "header") %>%
   italic(italic = TRUE, part = "header") %>%
   set_caption(
      "Table 3: Estimated powdery mildew severity mean difference ($u$) to the no spray control (intercept) for each spray schedule treatment. Estimates were calculated from a network meta-analysis of data obtained from grey literature reports of (k) field trials undertaken in Eastern Australia. P values indicate statistical significance in comparison to the intercept."
   ) %>%
   autofit() %>%
   footnote(
      i = 1,
      j = c(2:10),
      value = as_paragraph(
         c(
            "number of treatment means categorised to each spray schedule",
            "number of trials with the respective spray schedule",
            "estimated mean yield determined by the meta-analysis",
            "Lower range of the 95% confidence interval",
            "Upper range of the 95% confidence interval",
            "indicates the significance between each respective spray schedule and the no spray control (intercept)",
            "Fungicide spray schedule efficacy",
            "Lower range of the 95% confidence interval",
            "Upper range of the 95% confidence interval"
         )
      ),
      ref_symbols = letters[c(1:7,4,5)],
      part = "header",
      inline = TRUE
   )%>%
   add_header(Moderator = "",
              N = "",
              k = "",
              mu = "Effect Size",
              `CI_{L}` = "Effect Size",
              `CI_{U}` = "Effect Size",
              P  = "Effect Size",
              Efficacy = "PM control",
              `Eff CI_lower` = "PM control",
              `Eff CI upper` = "PM control") %>%
   merge_h( part = "header")%>%
   hline_top(part="all", border = officer::fp_border(color = "black", width = 2) )

disease_estimates_table
```

Save table to a word document to use in manuscript
```{r eval=FALSE, include=FALSE}
# Save table as a word document
doc <- officer::read_docx()
doc <- body_add_flextable(doc, value = disease_estimates_table)
print(doc, target = "paper/figures/Table3_severity_esimates.docx")
```
